codebase/requirements.txt
ib_insync
pandas
xlwings
schedule
numpy
nest_asyncio
pytz
pycryptodome